requirements.txt
fastapi
uvicorn
google-genai>=1.0.0
pydantic
python-dotenv
